Address code review findings
Copilot-Session: 4c723445-a61d-46cd-b02a-66d91f9004be
